ajustes no model n:n and serializer e endpoints

+++ api_rest/views.py
@@ -3,8 +3,8 @@ from rest_framework.decorators import api_view
 from rest_framework.response import Response
 from rest_framework import status
 
-from .models import Products, Sale, ProductSale
-from .serializer import ProductsSerializer, SaleSerializer, ProductSaleSerializer
+from .models import Products, Sale
+from .serializer import ProductsSerializer, SaleSerializer
 
 @api_view(['GET'])
 def get_products(request):
@@ -94,57 +94,50 @@ def create_sale(request):
 
 @api_view(['GET'])
 def list_sales(request):
-    sales = Sale.objects.all()
-    serializer = SaleSerializer(sales, many=True)
-    return Response(serializer.data)
-
+    if request.method == 'GET':
+        sales = Sale.objects.all()
+        serializer = SaleSerializer(sales, many=True)
+        return Response(serializer.data)
+    return Response(status=status.HTTP_400_BAD_REQUEST)
 
 @api_view(['GET'])
 def get_sale_by_id(request, sale_id):
     try:
         sale = Sale.objects.get(pk=sale_id)
-        product_sales = ProductSale.objects.filter(sale_id=sale)
-        product_sales_data = ProductSaleSerializer(product_sales, many=True).data
-        sale_data = SaleSerializer(sale).data
-        sale_data['products'] = product_sales_data
-        return Response(sale_data)
     except Sale.DoesNotExist:
         return Response(status=status.HTTP_404_NOT_FOUND)
 
+    if request.method == 'GET':
+        serializer = SaleSerializer(sale)
+        return Response(serializer.data)
+
+    return Response(status=status.HTTP_400_BAD_REQUEST)
 
 @api_view(['DELETE'])
 def cancel_sale(request, sale_id):
     try:
         sale = Sale.objects.get(pk=sale_id)
-        sale.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
     except Sale.DoesNotExist:
         return Response(status=status.HTTP_404_NOT_FOUND)
-        
 
+    if request.method == 'DELETE':
+        sale.delete()
+        return Response(status=status.HTTP_204_NO_CONTENT)
+
+    return Response(status=status.HTTP_400_BAD_REQUEST)
+        
 @api_view(['POST'])
 def add_products_to_sale(request, sale_id):
     try:
         sale = Sale.objects.get(pk=sale_id)
-        for item in request.data.get('product_sales', []):
-            ProductSale.objects.create(
-                product_id_id=item['product_id'],
-                sale_id=sale,
-                price=item['price'],
-                quantity=item['quantity']
-            )
-        return Response(status=status.HTTP_201_CREATED)
     except Sale.DoesNotExist:
         return Response(status=status.HTTP_404_NOT_FOUND)
-    if request.method == 'GET':
-        sales = Sale.objects.all()
-        serializer = SaleSerializer(sales, many=True)
-        return Response(serializer.data)
 
     if request.method == 'POST':
-        new_sale = request.data
-        serializer = SaleSerializer(data=new_sale)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data, status=status.HTTP_201_CREATED)
-        return Response(status=status.HTTP_400_BAD_REQUEST)
+        product_ids = request.data.get('product_ids', [])
+        products = Products.objects.filter(id__in=product_ids)
+
+        sale.Products.add(*products)
+
+        serializer = SaleSerializer(sale)
+        return Response(serializer.data, status=status.HTTP_200_OK)
